Modified extra principle and CSV output
So the CSV output has gotten uglier, but it means that it's a number
in a spreadsheet instead of text so you can highlight numbers to get
an average or sum of a particular range. Hopefully once I switch to
using the Decimal library it will become pretty again while retaining
usefulness.